server/requirements.txt
# Core Runtime Dependencies
fastapi>=0.95.0
uvicorn>=0.15.0
pydantic>=2.3.0
pydantic-settings==2.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
motor>=3.3.0
pymongo>=4.0,<5.0
python-dotenv>=0.21.0
google-generativeai>=0.3.0
python-multipart>=0.0.5,<0.1.0
httpx==0.26.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
# Added for Resume Parsing
pypdf>=4.0.0
python-docx>=1.1.0 # For handling .docx files
aiofiles>=23.1.0
python-magic>=0.4.27
